[ntuple,daos] Generalize RNTuple-Daos key mappings
Introduces `RDaosKey` to more flexibly switch between different mappings of NTuple metadata to DAOS location, along with the enumerator `EDaosMapping`. The function `GetDaosPageKey` determines a DAOS location according to the metadata as indicated by the templated argument of type `EDaosMapping`. The corresponding `RDaosKey` is generated and used in RW calls throughout `RPageStorageDaos`.